requirements copy.txt
Authlib==1.3.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-OAuthlib==0.9.6
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.28.1
oauthlib==2.1.0
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
requests-oauthlib==1.1.0
urllib3==1.26.16
Werkzeug==3.0.3
zipp==3.19.2
python-dotenv==1.0.0
boto3